Pipelined datapuller (#1016)
* Preliminary batching (untested)

* Prevent class deletion when 0 classes are pulled

* Insert classes with mongodb transactions

* Batch by termId + move deletion into mongodb transaction

* Preliminary batching (untested)

* Insert classes with mongodb transactions

* Batch by termId + move deletion into mongodb transaction

* Abort class datapuller if <90% of data is fetched

* Update fault tolerance of datapullers

Use mongodb transactions, aborting if we insert less than 95% of the
previous data.

Add more log messages to show how much data was deleted and inserted.

---------

+++ apps/datapuller/src/pullers/courses.ts
@@ -21,6 +21,15 @@ const updateCourses = async (config: Config) => {
 
   log.trace("Deleting courses no longer in SIS...");
 
+  const previousCourses = await CourseModel.countDocuments();
+
+  if (courses.length / previousCourses <= 0.95) {
+    log.error(
+      `Fetched only ${courses.length} courses, while there were ${previousCourses} previous courses`
+    );
+    return;
+  }
+
   const { deletedCount } = await CourseModel.deleteMany({
     courseId: { $nin: courses.map((course) => course.courseId) },
   });
